+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Bug fixes
 
 * Fixed issue in `find_parameters()` and `clean_parameters()` for *brmsfit* models with specific variable name patterns.
+* Fixed issue in `format_ci()` when confidence interval only contained `NA`s and `width` was set to `"auto"`.
 
 # insight 0.8.2