+++ NEWS.md
@@ -19,6 +19,7 @@
 * Fixed issue in `find_parameters()` and `clean_parameters()` for *brmsfit* models with specific variable name patterns.
 * Fixed issue in `format_ci()` when confidence interval only contained `NA`s and `width` was set to `"auto"`.
 * Fixed issue in `find_formula()` for mixed models when formula contained parentheses in the non-random parts, around a certain set of predictors.
+* Fixed issue in `get_priors.BFBayesFactor()` for `BFMetat` class.
 
 # insight 0.8.2